Add macros, functions and layouts to common.clj.
The two added macros are with-javascripts and with-title.  They allow
javascripts and site-title to be dynamically bindable for any given
page. The two layouts are `base-layout` and `layout`. The latter depends
on the former. These layouts are bootstrap friendly. Also, a helper
function, `alert` is added that displays a bootstrap friendly alert
message.

Most of this we copied from
[Cadence](https://github.com/RyanMcG/Cadence).